--- a/csv/src/main/mima-filters/1.0.x.backwards.excludes
+++ b/csv/src/main/mima-filters/1.0.x.backwards.excludes
@@ -1,0 +1,2 @@
+# Allow changes to impl
+ProblemFilters.exclude[Problem]("akka.stream.alpakka.csv.impl.*")