<!--
# Pull Request Checklist

* [x] Have you read through the [contributor guidelines](./CONTRIBUTING.md)?
* [x] Have you signed the [Lightbend CLA](https://www.lightbend.com/contribute/cla)?
* [x] Have you updated the documentation?
* [x] Have you added tests for any changed functionality?
-->
## Purpose

Improves parsing speed of alpakka-csv from 1.5 MiB/s to 78 MiB/s for the most extreme case.

## References

I searched, but didn't find any other issues or discussions.

## Changes

* New jmh benchmark: `CsvBench`
* `CsvParser` is faster.

## Background Context

While evaluating alpakka-csv as a non-blocking replacement for our currently blocking usage of OpenCSV, I benched it and found that it parses slower than I'd expect, even over benchmarks that play to alpakka-csv's strengths. After poking around, I found some slow spots and optimized them.

#### Before
```
[info] Benchmark       (bsSize)   Mode  Cnt    Score   Error  Units
[info] CsvBench.parse        32  thrpt   10    1.551 ± 0.028  ops/s
[info] CsvBench.parse      1024  thrpt   10   60.279 ± 0.720  ops/s
[info] CsvBench.parse      8192  thrpt   10  117.125 ± 1.138  ops/s
[info] CsvBench.parse     65536  thrpt   10  137.510 ± 3.048  ops/s
```

#### After
```
[info] Benchmark        (bsSize)   Mode  Cnt    Score   Error  Units
[info] CsvBench.parse        32  thrpt   10   78.424 ± 1.351  ops/s
[info] CsvBench.parse      1024  thrpt   10  158.948 ± 3.187  ops/s
[info] CsvBench.parse      8192  thrpt   10  167.617 ± 2.759  ops/s
[info] CsvBench.parse     65536  thrpt   10  170.670 ± 2.462  ops/s
```

### Slow spot 1: Duplicate parsing
On every new input `ByteString`, `parseLine: Option[List[ByteString]]` attempts to parse an entire row. If it doesn't have enough data to complete the row, it rewinds all progress on that row and waits for the next chunk then tries again. This causes duplicate work on every new chunk that continues an existing line. It gets particularly bad when the chunk size is significantly smaller than the row size.

There was already some state in `CsvParser` fields, so I pulled up the method state to class fields as well. It's fully incremental now and doesn't repeat work.

### Slow spot 2: `ByteStrings`/`MultiByteArrayIterator`
Accessing each byte and doing slices is hot inner-loop code and the composites (`ByteStrings`/`MultiByteArrayIterator`) add significant overhead to that.

We can't avoid `ByteStrings.slice(...)` when a field spans multiple chunks, but we can mitigate subsequent slices by calling `dropReadBuffer()` as soon as we complete a field, rather than waiting until we complete the whole line. It should allow `buffer: ByteString` to return to a fast `ByteString1` for all subsequent slices on the row.

Indexing into `ByteStrings.apply(pos)` for each byte is also costly. Even the `MultiByteArrayIterator` produced by `ByteStrings.iterator()` has to do non-free normalization on each `.next(): Byte` call.

An optimization is possible from the observation that `CsvParser` parses each `ByteString` chunk before receiving the next chunk. We can traverse only the single chunk in `parseLine` while keeping the `buffer: ByteStrings` around for slicing which may also need to contain the previous chunk.

Downside is we have to track a bit more state which adds a bit more complexity to the implementation. Upside is at least a 25% improvement across all scenarios, which makes it competitive with OpenCsv.