@ennru 2019-05-09
Thank you!

Great OSS attitude to dive into it when evaluating!

I was aware of slow spot 1 and tried to solve it earlier, but didn't have the time.