@htmldoug 2019-05-10
I tried `@inline` in a few places, but didn't see any improvement beyond the margin of error. In some cases, javap showed that scalac didn't actually inline the methods at all.

Top suggestion from JITWatch was:
```
The call at bytecode 1289 to
Class: akka.stream.alpakka.csv.impl.CsvParser
Member: dropReadBuffer()
was not inlined for reason: 'too big'
The callee method is not 'hot' but is too big to be inlined into the caller method.
Invocations: 8924
Size of callee bytecode: 39
```

One `dropReadBuffer()` stack frame for every csv field doesn't seem terrible. At least it's less frequent than every byte. Here's as much of the compile chain from JITWatch as I could capture on a single monitor: https://user-images.githubusercontent.com/663139/57494232-4c2a1d00-7296-11e9-82c9-47a7b758757f.png

I also tried extracting parts of `churn()` out into smaller methods to give the JIT more freedom to inline selectively, but perf actually got worse.