@htmldoug 2019-05-10
Thank you, @ennru! The fast review and helpful feedback makes contributing patches fun.