@ennru 2019-05-10
You can expect a release with the fastest Alpakka CSV ever next week.